# frontend/swagger-gen/swagger.json
{"swagger":"2.0","info":{"description":"Api Documentation","version":"1.0","title":"Api Documentation","termsOfService":"urn:tos","contact":{},"license":{"name":"Apache 2.0","url":"http://www.apache.org/licenses/LICENSE-2.0"}},"host":"localhost:8080","basePath":"/","tags":[{"name":"account-controller","description":"Account Controller"},{"name":"admin-controller","description":"Admin Controller"},{"name":"event-controller","description":"Event Controller"},{"name":"place-controller","description":"Place Controller"},{"name":"review-controller","description":"Review Controller"},{"name":"user-controller","description":"User Controller"}],"paths":{"/account/find-by-id/{id}":{"get":{"tags":["account-controller"],"summary":"Service that return a Account","description":"This service return a Account by the ID","operationId":"findById","produces":["application/json"],"parameters":[{"name":"Authorization","in":"header","required":true,"type":"string","x-example":"Bearer access_token"},{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"The request has succeeded.","schema":{"$ref":"#/definitions/AccountDTO"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not found"},"500":{"description":"Internal server error.","schema":{"$ref":"#/definitions/AccountDTO"}}}}},"/account/save":{"post":{"tags":["account-controller"],"summary":"This service save a Account","description":"Service that return AccountDTO with saved object Account","operationId":"save","consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"accountVO","description":"accountVO","required":true,"schema":{"$ref":"#/definitions/AccountVO"}},{"name":"Authorization","in":"header","required":true,"type":"string","x-example":"Bearer access_token"}],"responses":{"200":{"description":"The request has succeeded.","schema":{"$ref":"#/definitions/AccountDTO"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not found"},"500":{"description":"Internal server error.","schema":{"$ref":"#/definitions/AccountDTO"}}}}},"/create":{"post":{"tags":["user-controller"],"summary":"Service that return an user with details","description":"This service returns an user with details","operationId":"create","consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"user","description":"user","required":true,"schema":{"$ref":"#/definitions/UserVO"}}],"responses":{"200":{"description":"The request has succeeded.","schema":{"$ref":"#/definitions/UserDTO"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not found"},"500":{"description":"Internal server error.","schema":{"$ref":"#/definitions/UserDTO"}}}}},"/delete/{userId}/{id}":{"delete":{"tags":["admin-controller"],"summary":"This service delete a User","description":"Delete a User, if it doesn't find it throw an exception","operationId":"deleteById","produces":["application/json"],"parameters":[{"name":"Authorization","in":"header","required":true,"type":"string","x-example":"Bearer access_token"},{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"},{"name":"userId","in":"path","description":"userId","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"The request has succeeded.","schema":{"$ref":"#/definitions/BasicResponse"}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not found"},"500":{"description":"Internal server error.","schema":{"$ref":"#/definitions/BasicResponse"}}}}},"/event/delete/{accountId}/{eventId}":{"delete":{"tags":["event-controller"],"summary":"This service delete a event","description":"Delete a event, if it doesn't find it throw an exception","operationId":"removeFavorite","produces":["application/json"],"parameters":[{"name":"accountId","in":"path","description":"accountId","required":true,"type":"integer","format":"int64"},{"name":"Authorization","in":"header","required":true,"type":"string","x-example":"Bearer access_token"},{"name":"eventId","in":"path","description":"eventId","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"The request has succeeded.","schema":{"$ref":"#/definitions/BasicResponse"}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not found"},"500":{"description":"Internal server error.","schema":{"$ref":"#/definitions/BasicResponse"}}}}},"/event/favorite/{accountId}/{eventId}":{"put":{"tags":["event-controller"],"summary":"This service update a Event","description":"Update a Event, if it doesn't find it throw an exception","operationId":"favorite","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"accountId","in":"path","description":"accountId","required":true,"type":"integer","format":"int64"},{"name":"Authorization","in":"header","required":true,"type":"string","x-example":"Bearer access_token"},{"name":"eventId","in":"path","description":"eventId","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"The request has succeeded.","schema":{"$ref":"#/definitions/EventDTO"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not found"},"500":{"description":"Internal server error.","schema":{"$ref":"#/definitions/EventDTO"}}}}},"/event/find-all":{"get":{"tags":["event-controller"],"summary":"Service that returns all events","description":"This service returns all events load","operationId":"findAll_1","produces":["application/json"],"parameters":[{"name":"Authorization","in":"header","required":true,"type":"string","x-example":"Bearer access_token"}],"responses":{"200":{"description":"The request has succeeded.","schema":{"type":"array","items":{"$ref":"#/definitions/EventDTO"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not found"},"500":{"description":"Internal server error.","schema":{"type":"array","items":{"$ref":"#/definitions/EventDTO"}}}}}},"/event/find-all-by-user-id/{userId}":{"get":{"tags":["event-controller"],"summary":"Service that returns all events by user","description":"This service returns all events load by user","operationId":"findAllByUserId","produces":["application/json"],"parameters":[{"name":"Authorization","in":"header","required":true,"type":"string","x-example":"Bearer access_token"},{"name":"userId","in":"path","description":"userId","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"The request has succeeded.","schema":{"type":"array","items":{"$ref":"#/definitions/EventDTO"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not found"},"500":{"description":"Internal server error.","schema":{"type":"array","items":{"$ref":"#/definitions/EventDTO"}}}}}},"/event/find-by-id/{id}":{"get":{"tags":["event-controller"],"summary":"Service that return a Event","description":"This service return a Event by the ID","operationId":"findById_2","produces":["application/json"],"parameters":[{"name":"Authorization","in":"header","required":true,"type":"string","x-example":"Bearer access_token"},{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"The request has succeeded.","schema":{"$ref":"#/definitions/EventDTO"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not found"},"500":{"description":"Internal server error.","schema":{"$ref":"#/definitions/EventDTO"}}}}},"/event/save/{userId}/{eventId}":{"post":{"tags":["event-controller"],"summary":"This service save a Event","description":"Service that return EventDTO with saved object Event","operationId":"save_1","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"Authorization","in":"header","required":true,"type":"string","x-example":"Bearer access_token"},{"name":"eventId","in":"path","description":"eventId","required":true,"type":"integer","format":"int64"},{"in":"body","name":"eventVO","description":"eventVO","required":true,"schema":{"$ref":"#/definitions/EventVO"}},{"name":"userId","in":"path","description":"userId","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"The request has succeeded.","schema":{"$ref":"#/definitions/EventDTO"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not found"},"500":{"description":"Internal server error.","schema":{"$ref":"#/definitions/EventDTO"}}}}},"/event/total-review-score/{eventId}":{"get":{"tags":["event-controller"],"summary":"Service that returns all value score","description":"This service returns all value score","operationId":"getTotalReviewScore","produces":["application/json"],"parameters":[{"name":"Authorization","in":"header","required":true,"type":"string","x-example":"Bearer access_token"},{"name":"eventId","in":"path","description":"eventId","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"The requesvt has succeeded.","schema":{"type":"number"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not found"},"500":{"description":"Internal server error.","schema":{"type":"number"}}}}},"/event/update/{userId}/{eventId}":{"put":{"tags":["event-controller"],"summary":"This service update a Event","description":"Update a Event, if it doesn't find it throw an exception","operationId":"update_1","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"Authorization","in":"header","required":true,"type":"string","x-example":"Bearer access_token"},{"name":"eventId","in":"path","description":"eventId","required":true,"type":"integer","format":"int64"},{"in":"body","name":"review","description":"review","required":true,"schema":{"$ref":"#/definitions/ReviewVO"}},{"name":"userId","in":"path","description":"userId","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"The request has succeeded.","schema":{"$ref":"#/definitions/EventDTO"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not found"},"500":{"description":"Internal server error.","schema":{"$ref":"#/definitions/EventDTO"}}}}},"/find-all":{"get":{"tags":["admin-controller"],"summary":"Service that returns all users","description":"This service returns all users load","operationId":"findAll","produces":["application/json"],"parameters":[{"name":"Authorization","in":"header","required":true,"type":"string","x-example":"Bearer access_token"}],"responses":{"200":{"description":"The request has succeeded.","schema":{"type":"array","items":{"$ref":"#/definitions/UserDTO"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not found"},"500":{"description":"Internal server error.","schema":{"type":"array","items":{"$ref":"#/definitions/UserDTO"}}}}}},"/find-by-id/{id}":{"get":{"tags":["admin-controller"],"summary":"Service that return a User","description":"This service return a User by the ID","operationId":"findById_1","produces":["application/json"],"parameters":[{"name":"Authorization","in":"header","required":true,"type":"string","x-example":"Bearer access_token"},{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"The request has succeeded.","schema":{"$ref":"#/definitions/UserDTO"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not found"},"500":{"description":"Internal server error.","schema":{"$ref":"#/definitions/UserDTO"}}}}},"/login":{"post":{"tags":["user-controller"],"summary":"Service that return an user with token access","description":"This service returns an user with token access","operationId":"login","consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"user","description":"user","required":true,"schema":{"$ref":"#/definitions/UserLoginVO"}}],"responses":{"200":{"description":"The request has succeeded.","schema":{"$ref":"#/definitions/JwtResponseDTO"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not found"},"500":{"description":"Internal server error.","schema":{"$ref":"#/definitions/JwtResponseDTO"}}}}},"/place/delete/{accountId}/{placeId}":{"delete":{"tags":["place-controller"],"summary":"This service delete a place","description":"Delete a place, if it doesn't find it throw an exception","operationId":"removeFavorite_1","produces":["application/json"],"parameters":[{"name":"accountId","in":"path","description":"accountId","required":true,"type":"integer","format":"int64"},{"name":"Authorization","in":"header","required":true,"type":"string","x-example":"Bearer access_token"},{"name":"placeId","in":"path","description":"placeId","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"The request has succeeded.","schema":{"$ref":"#/definitions/BasicResponse"}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not found"},"500":{"description":"Internal server error.","schema":{"$ref":"#/definitions/BasicResponse"}}}}},"/place/favorite/{accountId}/{placeId}":{"put":{"tags":["place-controller"],"summary":"This service update a Place","description":"Update a Place, if it doesn't find it throw an exception","operationId":"favorite_1","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"accountId","in":"path","description":"accountId","required":true,"type":"integer","format":"int64"},{"name":"Authorization","in":"header","required":true,"type":"string","x-example":"Bearer access_token"},{"name":"placeId","in":"path","description":"placeId","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"The request has succeeded.","schema":{"$ref":"#/definitions/PlaceDTO"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not found"},"500":{"description":"Internal server error.","schema":{"$ref":"#/definitions/PlaceDTO"}}}}},"/place/find-all":{"get":{"tags":["place-controller"],"summary":"Service that returns all places","description":"This service returns all places load","operationId":"findAll_2","produces":["application/json"],"parameters":[{"name":"Authorization","in":"header","required":true,"type":"string","x-example":"Bearer access_token"}],"responses":{"200":{"description":"The request has succeeded.","schema":{"type":"array","items":{"$ref":"#/definitions/PlaceDTO"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not found"},"500":{"description":"Internal server error.","schema":{"type":"array","items":{"$ref":"#/definitions/PlaceDTO"}}}}}},"/place/find-all-by-user-id/{userId}":{"get":{"tags":["place-controller"],"summary":"Service that returns all places by user","description":"This service returns all places load by user","operationId":"findAllByUserId_1","produces":["application/json"],"parameters":[{"name":"Authorization","in":"header","required":true,"type":"string","x-example":"Bearer access_token"},{"name":"userId","in":"path","description":"userId","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"The request has succeeded.","schema":{"type":"array","items":{"$ref":"#/definitions/PlaceDTO"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not found"},"500":{"description":"Internal server error.","schema":{"type":"array","items":{"$ref":"#/definitions/PlaceDTO"}}}}}},"/place/find-by-id/{id}":{"get":{"tags":["place-controller"],"summary":"Service that return a Place","description":"This service return a Place by the ID","operationId":"findById_3","produces":["application/json"],"parameters":[{"name":"Authorization","in":"header","required":true,"type":"string","x-example":"Bearer access_token"},{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"The request has succeeded.","schema":{"$ref":"#/definitions/PlaceDTO"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not found"},"500":{"description":"Internal server error.","schema":{"$ref":"#/definitions/PlaceDTO"}}}}},"/place/save/{userId}/{placeId}":{"post":{"tags":["place-controller"],"summary":"This service save a Place","description":"Service that return PlaceDTO with saved object Place","operationId":"save_2","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"Authorization","in":"header","required":true,"type":"string","x-example":"Bearer access_token"},{"name":"placeId","in":"path","description":"placeId","required":true,"type":"integer","format":"int64"},{"in":"body","name":"placeVO","description":"placeVO","required":true,"schema":{"$ref":"#/definitions/PlaceVO"}},{"name":"userId","in":"path","description":"userId","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"The request has succeeded.","schema":{"$ref":"#/definitions/PlaceDTO"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not found"},"500":{"description":"Internal server error.","schema":{"$ref":"#/definitions/PlaceDTO"}}}}},"/place/total-review-score/{placeId}":{"get":{"tags":["place-controller"],"summary":"Service that returns all value score","description":"This service returns all value score","operationId":"getTotalReviewScore_1","produces":["application/json"],"parameters":[{"name":"Authorization","in":"header","required":true,"type":"string","x-example":"Bearer access_token"},{"name":"placeId","in":"path","description":"placeId","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"The requesvt has succeeded.","schema":{"type":"number"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not found"},"500":{"description":"Internal server error.","schema":{"type":"number"}}}}},"/place/update/{userId}/{placeId}":{"put":{"tags":["place-controller"],"summary":"This service update a Place","description":"Update a Place, if it doesn't find it throw an exception","operationId":"update_2","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"Authorization","in":"header","required":true,"type":"string","x-example":"Bearer access_token"},{"name":"placeId","in":"path","description":"placeId","required":true,"type":"integer","format":"int64"},{"in":"body","name":"review","description":"review","required":true,"schema":{"$ref":"#/definitions/ReviewVO"}},{"name":"userId","in":"path","description":"userId","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"The request has succeeded.","schema":{"$ref":"#/definitions/PlaceDTO"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not found"},"500":{"description":"Internal server error.","schema":{"$ref":"#/definitions/PlaceDTO"}}}}},"/review/find-all":{"get":{"tags":["review-controller"],"summary":"Service that returns all reviews","description":"This service returns all reviews load","operationId":"findAll_3","produces":["application/json"],"parameters":[{"name":"Authorization","in":"header","required":true,"type":"string","x-example":"Bearer access_token"}],"responses":{"200":{"description":"The request has succeeded.","schema":{"type":"array","items":{"$ref":"#/definitions/AccountReviewDetailsDTO"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not found"},"500":{"description":"Internal server error.","schema":{"type":"array","items":{"$ref":"#/definitions/AccountReviewDetailsDTO"}}}}}},"/review/find-by-id/{id}":{"get":{"tags":["review-controller"],"summary":"Service that return a Review","description":"This service return a Review by the ID","operationId":"findById_4","produces":["application/json"],"parameters":[{"name":"Authorization","in":"header","required":true,"type":"string","x-example":"Bearer access_token"},{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"The request has succeeded.","schema":{"$ref":"#/definitions/ReviewDTO"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not found"},"500":{"description":"Internal server error.","schema":{"$ref":"#/definitions/ReviewDTO"}}}}},"/review/get-review/{placeId}/{accountId}":{"get":{"tags":["review-controller"],"summary":"Service that return a Review","description":"This service return a Review by the ID","operationId":"getReview","produces":["application/json"],"parameters":[{"name":"accountId","in":"path","description":"accountId","required":true,"type":"integer","format":"int64"},{"name":"Authorization","in":"header","required":true,"type":"string","x-example":"Bearer access_token"},{"name":"placeId","in":"path","description":"placeId","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"The request has succeeded.","schema":{"$ref":"#/definitions/ReviewDTO"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not found"},"500":{"description":"Internal server error.","schema":{"$ref":"#/definitions/ReviewDTO"}}}}},"/update/{userId}":{"put":{"tags":["admin-controller"],"summary":"This service update a User","description":"Update a User, if it doesn't find it throw an exception","operationId":"update","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"Authorization","in":"header","required":true,"type":"string","x-example":"Bearer access_token"},{"in":"body","name":"user","description":"user","required":true,"schema":{"$ref":"#/definitions/UserDTO"}},{"name":"userId","in":"path","description":"userId","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"The request has succeeded.","schema":{"$ref":"#/definitions/UserDTO"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not found"},"500":{"description":"Internal server error.","schema":{"$ref":"#/definitions/UserDTO"}}}}}},"definitions":{"AccountDTO":{"type":"object","properties":{"address":{"type":"string"},"dni":{"type":"string"},"firstname":{"type":"string"},"id":{"type":"integer","format":"int64"},"imagePath":{"type":"string"},"lastname":{"type":"string"},"phoneNumber":{"type":"string"},"role":{"type":"string"},"roleDescripton":{"type":"string"}},"title":"AccountDTO"},"AccountReviewDetailsDTO":{"type":"object","properties":{"accountId":{"type":"integer","format":"int64"},"comments":{"type":"string"},"dni":{"type":"string"},"eventId":{"type":"integer","format":"int64"},"eventType":{"type":"string"},"eventTypeDescription":{"type":"string"},"firstname":{"type":"string"},"lastname":{"type":"string"},"placeId":{"type":"integer","format":"int64"},"placeName":{"type":"string"},"placeType":{"type":"string"},"placeTypeDescription":{"type":"string"},"reviewId":{"type":"integer","format":"int64"},"score":{"type":"number","format":"double"},"title":{"type":"string"}},"title":"AccountReviewDetailsDTO"},"AccountVO":{"type":"object","properties":{"address":{"type":"string"},"dni":{"type":"string"},"firstname":{"type":"string"},"imagePath":{"type":"string"},"lastname":{"type":"string"},"phoneNumber":{"type":"string"},"role":{"type":"string"}},"title":"AccountVO"},"AuthorityDTO":{"type":"object","properties":{"email":{"type":"string"},"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"position":{"type":"string"}},"title":"AuthorityDTO"},"BasicResponse":{"type":"object","properties":{"error":{"type":"boolean"},"message":{"type":"string"}},"title":"BasicResponse"},"EventDTO":{"type":"object","properties":{"body":{"type":"string"},"dependsOn":{"type":"string"},"email":{"type":"string"},"eventId":{"type":"integer","format":"int64"},"eventType":{"type":"string"},"eventTypeDescription":{"type":"string"},"fromDate":{"type":"string"},"id":{"type":"integer","format":"int64"},"image":{"type":"string"},"isFavorite":{"type":"boolean"},"link":{"type":"string"},"state":{"type":"string"},"subTitle":{"type":"string"},"title":{"type":"string"},"toDate":{"type":"string"},"url":{"type":"string"}},"title":"EventDTO"},"EventVO":{"type":"object","properties":{"body":{"type":"string"},"dependsOn":{"type":"string"},"email":{"type":"string"},"eventId":{"type":"integer","format":"int64"},"eventType":{"type":"string"},"fromDate":{"type":"string"},"image":{"type":"string"},"link":{"type":"string"},"state":{"type":"string"},"subTitle":{"type":"string"},"title":{"type":"string"},"toDate":{"type":"string"},"url":{"type":"string"}},"title":"EventVO"},"JwtResponseDTO":{"type":"object","properties":{"accountId":{"type":"integer","format":"int64"},"email":{"type":"string"},"firstname":{"type":"string"},"lastname":{"type":"string"},"role":{"type":"string"},"token":{"type":"string"},"userId":{"type":"integer","format":"int64"}},"title":"JwtResponseDTO"},"PlaceDTO":{"type":"object","properties":{"address":{"type":"string"},"authority":{"$ref":"#/definitions/AuthorityDTO"},"dependsOn":{"type":"string"},"description":{"type":"string"},"email":{"type":"string"},"id":{"type":"integer","format":"int64"},"isFavorite":{"type":"boolean"},"link":{"type":"string"},"name":{"type":"string"},"phoneNumber":{"type":"string"},"placeId":{"type":"integer","format":"int64"},"placeType":{"type":"string"},"placeTypeDescription":{"type":"string"},"province":{"type":"string"},"url":{"type":"string"}},"title":"PlaceDTO"},"PlaceVO":{"type":"object","properties":{"address":{"type":"string"},"authorityEmail":{"type":"string"},"authorityName":{"type":"string"},"authorityPosition":{"type":"string"},"dependsOn":{"type":"string"},"description":{"type":"string"},"email":{"type":"string"},"link":{"type":"string"},"name":{"type":"string"},"phoneNumber":{"type":"string"},"placeId":{"type":"integer","format":"int64"},"placeType":{"type":"string"},"placeTypeDescription":{"type":"string"},"province":{"type":"string"},"url":{"type":"string"}},"title":"PlaceVO"},"ReviewDTO":{"type":"object","properties":{"comments":{"type":"string"},"id":{"type":"integer","format":"int64"},"score":{"type":"number","format":"double"}},"title":"ReviewDTO"},"ReviewVO":{"type":"object","properties":{"comments":{"type":"string"},"score":{"type":"number","format":"double"}},"title":"ReviewVO"},"UserDTO":{"type":"object","properties":{"account":{"$ref":"#/definitions/AccountDTO"},"email":{"type":"string"},"id":{"type":"integer","format":"int64"},"password":{"type":"string"}},"title":"UserDTO"},"UserLoginVO":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"title":"UserLoginVO"},"UserVO":{"type":"object","properties":{"account":{"$ref":"#/definitions/AccountVO"},"email":{"type":"string"},"password":{"type":"string"},"repeatPassword":{"type":"string"}},"title":"UserVO"}}}